+++ Truck_Out.cpp
@@ -9,7 +9,7 @@ void Out(Truck* t, ofstream& ofst)
 	ofst << "It is Truck, carrying capacity = " << t->mData << ", Engine power = " << t->tPower << endl;
 };
 
-int WPRatio(Truck* t)
+float WPRatio(Truck* t)
 {
-	return t->mData/t->tPower;
+	return (float)t->mData/(float)t->tPower;
 };
